[FIX] fixed the commented changes on PR
