+++ sale_backorder/models/__init__.py
@@ -1,3 +1,4 @@
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
 from . import sale
+from . import sale_order_line
